+++ HISTORY.rst
@@ -10,6 +10,7 @@ History
   ``article/front/article-meta/aff/institution/@content-type``.
 * Clean up unused attributes from classes from the ``packtools.style_error`` 
   module. 
+* Log messages are now omitted from the stderr by default. 
 
 
 1.3.2 (2016-11-22)